refactor: ResultListItem 상단 변경되는 텍스트와 하단 도움말 박스 컴포넌트 분리
